% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema.array}
\alias{schema.array}
\title{scidb base functions}
\usage{
schema.array(a_name)
}
\arguments{
\item{a_name}{A valid array name. If \code{NULL} all arrays are returned (Default \code{NULL}).}
}
\value{
A schema string
}
\description{
Generate a new schema string
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
