% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema.dim}
\alias{schema.dim}
\title{scidb base functions}
\usage{
schema.dim(schema)
}
\arguments{
\item{schema}{A scidb schema string.}
}
\value{
A list with all array dimensions and its range and chunck propoerties.
}
\description{
Generate a new schema string
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
