% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.op.R
\name{scidb.apply}
\alias{scidb.apply}
\title{scidb operators functions}
\usage{
scidb.apply(a, ...)
}
\arguments{
\item{a}{A valid array name returned by \code{\link{scidb.array}} 
or derived AFL command.}

\item{...}{Pairs of \code{attr_name=value} arguments, where \code{value} 
can be any valid scidb expression.}
}
\value{
AFL string command
}
\description{
This function wraps the scidb \code{"apply"} operator. 
              The operator creates new attributes to be returned.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
