% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{scidb.array}
\alias{scidb.array}
\title{scidb base functions}
\usage{
scidb.array(a_name)
}
\arguments{
\item{a_name}{A valid array name. If \code{NULL} all arrays are returned (Default \code{NULL})}
}
\value{
AFL string command
}
\description{
Get a specific scidb array
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
