% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{scidb.array_schema}
\alias{scidb.array_schema}
\title{scidb base functions}
\usage{
scidb.array_schema(a_name, R_types = FALSE)
}
\arguments{
\item{a_name}{A valid array name. If \code{NULL} all arrays are returned (Default \code{NULL}).}

\item{R_types}{Logical. Convert scidb types to compatible R types (Default \code{FALSE}).}
}
\value{
AFL string command
}
\description{
Get a specific scidb array schema
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
