% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{scidb.arrays}
\alias{scidb.arrays}
\title{scidb base functions}
\usage{
scidb.arrays()
}
\value{
Tibble with AFL result
}
\description{
Get all scidb arrays
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
