% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.op.R
\name{scidb.between}
\alias{scidb.between}
\title{scidb operators functions}
\usage{
scidb.between(a, ..., max_len)
}
\arguments{
\item{a}{A valid array name returned by \code{\link{scidb.array}} 
or derived AFL command.}

\item{...}{Pairs of \code{dim_name=c(min, max)} arguments.}

\item{max_len}{The max length accepted and used to break apart the informed dimensions ranges
into small pieces of ranges.}
}
\value{
AFL string command
}
\description{
This function wraps the scidb \code{"between"} operator. 
              The operator filter dimensions range to be returned.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
