% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{scidb.cast}
\alias{scidb.cast}
\title{scidb operators functions}
\usage{
scidb.cast(a, schema)
}
\arguments{
\item{a}{A valid array name or derived AFL command.}

\item{schema}{A valid scidb schema string. Can be generated by \code{scidb.make_schema} 
function or by \code{scidb.array.schema}. The content syntax is not verified.}
}
\value{
AFL string command
}
\description{
This function wraps the scidb \code{"cast"} operator. 
              The operator casts the attributes types to be returned.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
