% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\docType{data}
\name{scidb.env}
\alias{scidb.env}
\title{scidb global variable}
\format{An object of class \code{environment} of length 1.}
\usage{
scidb.env
}
\description{
Package environment
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{datasets}
