% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{scidb.iquery_path}
\alias{scidb.iquery_path}
\title{scidb base functions}
\usage{
scidb.iquery_path()
}
\value{
Path to iquery binary
}
\description{
Get actual path of iquery binary
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
