% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{scidb.list}
\alias{scidb.list}
\title{scidb base functions}
\usage{
scidb.list(element = "arrays")
}
\arguments{
\item{element}{Type of entity to be listed, one of:
\code{"aggregates"}, \code{"arrays"}, \code{"functions"}, \code{"instances"}, 
\code{"libraries"}, \code{"operators"}, \code{"types"}, and \code{"queries"}
(Default \code{"arrays"}).}
}
\value{
Tibble with AFL result
}
\description{
Get all scidb arrays
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
