% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{scidb.make_schema}
\alias{scidb.make_schema}
\title{scidb base functions}
\usage{
scidb.make_schema(a_name, sch_attrs = list(attr1 = "int32"),
  sch_dim = list(dim1 = c(0, 10, 11, 0)))
}
\arguments{
\item{a_name}{A valid array name. If \code{NULL} all arrays are returned (Default \code{NULL}).}

\item{sch_attrs}{A list of pairs \code{attr_name="type"}, where \code{"type"} is any scidb type.
R compatible types: \code{"int32"}, \code{"double"}, and \code{"string"}.}

\item{sch_dim}{A list of pairs \code{dim_name=c(from, to, chunck_size, chunk_overlap)}, where
\code{from}, \code{to}, \code{chunk_size}, and \code{chunk_overlap} are numbers.}
}
\value{
A schema string
}
\description{
Generate a new schema string
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
