% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{scidb.map_stream}
\alias{scidb.map_stream}
\title{scidb stream functions}
\usage{
scidb.map_stream(fun, err_file = NULL)
}
\arguments{
\item{fun}{A function to be called to process the input data which comes from scidb.
The function must have one argument to receive the chunck data.
The result of this function is returned to the scidb stream operator.}

\item{err_file}{A file path where the script can report any errors. If \code{NULL}, any
error will be reported on \code{stderr} (Default \code{NULL}).}
}
\value{
Named list with all decoded script arguments
}
\description{
This function get all script arguments, decodefies them and return a named list.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
