% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\docType{data}
\name{scidb.path}
\alias{scidb.path}
\title{scidb global variable}
\format{An object of class \code{character} of length 1.}
\usage{
scidb.path
}
\description{
Path to iquery binary
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{datasets}
