% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{scidb.redimension}
\alias{scidb.redimension}
\title{scidb operators functions}
\usage{
scidb.redimension(a, schema)
}
\arguments{
\item{a}{A valid array name or derived AFL command.}

\item{schema}{A valid scidb schema string. Can be generated by \code{scidb.make_schema} 
function or by \code{scidb.array.schema}. The content syntax is not verified.}
}
\value{
AFL string command
}
\description{
This function wraps the scidb \code{"redimension"} operator. 
              The operator changes the dimensions and attributes of the input array.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
