% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{scidb.Rscript_params}
\alias{scidb.Rscript_params}
\title{scidb stream functions}
\usage{
scidb.Rscript_params(required_params = NULL, err_file = NULL)
}
\arguments{
\item{required_params}{A list of pairs \code{param_name="type"} containing all required 
parameters and its respective type. The valid types are \code{"int32"},
\code{"double"}, and \code{"string"}. If any parameter
is missing or has a wrong type, an error is throwed and reported. 
If \code{NULL}, no parameter is required (Default \code{NULL}).}

\item{err_file}{A file path where the script can report any errors. If \code{NULL}, any
error will be reported on \code{stderr} only (Default \code{NULL}).}
}
\value{
Named list with all decoded script arguments
}
\description{
This function get all script arguments, decodefies them and return a named list.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
