% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{scidb.set_iquery_path}
\alias{scidb.set_iquery_path}
\title{scidb base functions}
\usage{
scidb.set_iquery_path(path)
}
\arguments{
\item{path}{Path to iquery binary}
}
\value{
String of path
}
\description{
Get actual path of iquery binary
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
