% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.op.R
\name{scidb.store}
\alias{scidb.store}
\title{scidb operators functions}
\usage{
scidb.store(a, a_name)
}
\arguments{
\item{a}{A valid array name returned by \code{\link{scidb.array}} 
or derived AFL command.}

\item{a_name}{A valid scidb array name string.}
}
\value{
AFL string command
}
\description{
This function wraps the scidb \code{"store"} operator. 
              The operator stores the resulting array into a new array name.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
