% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.op.R
\name{scidb.stream}
\alias{scidb.stream}
\title{scidb operators functions}
\usage{
scidb.stream(a, file, params = NULL, file_output_fields)
}
\arguments{
\item{a}{A valid array name returned by \code{\link{scidb.array}} 
or derived AFL command.}

\item{file}{A R file script to be executed.}

\item{params}{Optional. A list of pairs \code{name=value} to be passed as input arguments 
to the script. The \code{value} argument can be either numeric or string.
If a vector of those types is passed, each element will be separated by commas.}

\item{file_output_fields}{A list of pairs \code{field_name=value}, where \code{value} is any valid 
scidb types compatible with R, i.e. (\code{"int32"}, \code{"double"}, or \code{"string"}).}
}
\value{
AFL string command
}
\description{
This function wraps the scidb \code{"stream"} operator. 
              The operator calls a R script and pass as input an array.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
