% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\docType{data}
\name{scidb.valid_types}
\alias{scidb.valid_types}
\title{scidb global variable}
\format{An object of class \code{character} of length 3.}
\usage{
scidb.valid_types
}
\description{
Compatible scidb to R types
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{datasets}
