% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{scidb.exec}
\alias{scidb.exec}
\title{scidb base functions}
\usage{
scidb.exec(afl, fetch_data = TRUE)
}
\arguments{
\item{afl}{A valid AFL scidb query}

\item{fetch_data}{Logical. Inform if data must be retrieved from \code{afl} query (Default \code{TRUE}).}
}
\value{
Tibble with AFL result
}
\description{
Execute AFL instruction
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
