% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{scidb.stream}
\alias{scidb.stream}
\title{scidb operators functions}
\usage{
scidb.stream(a, script, params = NULL)
}
\arguments{
\item{a}{A valid array name or derived AFL command.}

\item{script}{A R file script to be executed.}

\item{params}{Optional. A list of pairs \code{name=value} to be passed as input arguments 
to the script. The \code{value} argument can be either numeric or string.
If a vector of those types is passed, each element will be separated by commas.}
}
\value{
AFL string command
}
\description{
This function wraps the scidb \code{"stream"} operator. 
              The operator calls a R script and pass as input an array.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
